+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 - fixed: Show a useful, localized error message when device doesn't have an email account
 - fixed: Make FilledTextInputs take up constant vertical space
 - fixed: Send Recipient Address Modal styling when saved recipients are shown
+- fixed: Minor Transaction List spacing adjustments
+- fixed: Insufficient Funds error in Thorchain unstaking
 - removed: swipeLastUsp experiment (always allow swipes on the last USP)
 
 ## 4.0.1 (2024-02-04)